requirements.txt
streamlit
pymongo
python-dotenv
bcrypt
pandas
matplotlib
numpy
openpyxl
requests
pillow
pyyaml
toml